backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
opencv-python-headless==4.10.0.84
numpy>=1.26.0
scikit-learn>=1.5.0
scipy>=1.10.0
Pillow>=10.4.0
motor>=3.4.0
pymongo[srv]>=4.7.0
python-dotenv>=1.0.0
